requirements.txt
mysql_connector
pyTelegramBotAPI